+++ packages/perseus/src/widgets/image/__docs__/explore-image-modal.stories.tsx
@@ -2,7 +2,14 @@ import * as React from "react";
 
 import {ApiOptions} from "../../../perseus-api";
 import {ExploreImageModal} from "../components/explore-image-modal";
-import {earthMoonImage, frescoImage, monasteryImage} from "../utils";
+import {
+    earthMoonImage,
+    frescoImage,
+    portraitImage,
+    portraitImageCaption,
+    portraitImageLongDescription,
+    portraitImageTitle,
+} from "../utils";
 
 import styles from "./image-stories.module.css";
 
@@ -61,13 +68,11 @@ export const SmallImage: Story = {
 
 export const PortraitImage: Story = {
     args: {
-        backgroundImage: monasteryImage,
-        zoomSize: [monasteryImage.width, monasteryImage.height] satisfies Size,
-        caption:
-            "Kalenić Monastery, after 1407, Serbia (photo: [Ванилица](https://commons.wikimedia.org/wiki/File:Wiki_%C5%A0umadija_XI_Kaleni%C4%87_Monastery_874.jpg), CC BY-SA 4.0)",
-        title: "Kalenić Monastery",
-        longDescription:
-            "Later architecture in Serbia, notably that of the so-called Morava School, is smaller and more decorative, often utilizing the so-called Athonite plan (with choroi and subsidiary chapels), as at Ravanica (1370s), with five domes, or the smaller and simpler Kalenić (after 1407).",
+        backgroundImage: portraitImage,
+        zoomSize: [portraitImage.width, portraitImage.height] satisfies Size,
+        caption: portraitImageCaption,
+        title: portraitImageTitle,
+        longDescription: portraitImageLongDescription,
         apiOptions: ApiOptions.defaults,
     },
 };